+++ src/Commands/ChangelogUnreleasedCommand.php
@@ -5,6 +5,7 @@ namespace MarkWalet\Changelog\Commands;
 use Illuminate\Console\Command;
 use MarkWalet\Changelog\Adapters\ReleaseAdapter;
 use MarkWalet\Changelog\ChangelogFormatterFactory;
+use MarkWalet\Changelog\Exceptions\InvalidXmlException;
 use MarkWalet\Changelog\Release;
 
 class ChangelogUnreleasedCommand extends Command
@@ -34,28 +35,21 @@ class ChangelogUnreleasedCommand extends Command
         $formatter = $factory->make('text');
         $path = config('changelog.path');
 
-        $release = $this->release($adapter, $path);
+
+        try {
+            $release = $adapter->exists($path, 'unreleased')
+                ? $adapter->read($path, 'unreleased')
+                : new Release('unreleased');
+        } catch(InvalidXmlException $e) {
+            $this->error('Something went wrong while parsing the xml file.');
+
+            return;
+        }
 
         $lines = explode(PHP_EOL, $formatter->single($release));
 
         foreach ($lines as $line) {
             $this->line($line);
         }
-    }
-
-    /**
-     * Get the unreleased feature.
-     *
-     * @param ReleaseAdapter $adapter
-     * @param string $path
-     * @return Release
-     */
-    private function release(ReleaseAdapter $adapter, string $path): Release
-    {
-        if ($adapter->exists($path, 'unreleased') === false) {
-            return new Release('unreleased');
-        }
-
-        return $adapter->read($path, 'unreleased');
     }
 }